Added inkscape extension ("-e inkscape") and test_inkscape test.
The extension needs some cleanup, but seems to be
a good proof-of-concept.  In any case, it works
well enough for me to use at work, so I might as
well put it with the other rstpdf code where I
won't lose it :-)

It allows import of .svg images by calling inkscape
to convert them into .pdf images, then using pdfrw
to read in the .pdf images.  Inkscape can handle
some .svg images that uniconvertor and svglib do
not do a very good job on.